backend/requirements.txt
fastapi==0.124.0
starlette==0.50.0
uvicorn==0.24.0

websockets==12.0

httpx==0.28.1
httpcore==1.0.9
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiosqlite==0.19.0
email-validator==2.1.0

langchain==1.1.3
langchain-core==1.1.3
langchain-community==0.4.1
langchain-openai==1.1.1
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-sdk==0.2.15
langsmith==0.4.58

openai==2.9.0
tiktoken==0.12.0

pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5

orjson==3.11.5
zstandard==0.25.0
xxhash==3.6.0
ormsgpack==1.12.0

requests==2.32.5
tenacity==9.1.2
PyYAML==6.0.3
regex==2025.11.3
typing_extensions==4.15.0
